fixup! chore: move multi-org endpoints into enterprise directory
